% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customplot.R
\name{customplot}
\alias{customplot}
\title{Plots a linear piece-wise regression}
\usage{
customplot(xk, dataframe, xvar, yvar)
}
\arguments{
\item{xk}{The x value to split the regression}

\item{dataframe}{The data frame to plot}

\item{xvar}{The name of the x var}

\item{yvar}{The name of the y var}
}
\value{
Returns a plot of the piece-wise regression
}
\description{
Plots a linear piece-wise regression
}
