% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Plots a histogram of the function with the confidence interval}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{The number of iterations to do}

\item{x}{The distribution to calculate}

\item{fun}{The function to apply to the distribution}

\item{alpha}{The confidence interval of the function}

\item{cx}{The numeric value of the plot symbol}

\item{...}{Additional parameters for the histogram}
}
\value{
The histogram of the confidence interval
}
\description{
Plots a histogram of the function with the confidence interval
}
