% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Plot the histogram of the sample mean of a binomial distribution}
\usage{
mycltb(n, iter, p = 0.5, ...)
}
\arguments{
\item{n}{The sample size}

\item{iter}{The number of iterations}

\item{p}{The probability of success}

\item{...}{Additional parameters for the histogram}
}
\description{
Plot the histogram of the sample mean of a binomial distribution
}
