% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Creates a norm curve with a probability region}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean for the distribution}

\item{sigma}{The standard deviation of the distribution}

\item{a}{The number to calculate probability using}
}
\value{
Plots a norm curve with a probability region
}
\description{
Creates a norm curve with a probability region
}
