% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynrml.R
\name{myNRML}
\alias{myNRML}
\title{Creates a NR maximum likelihood estimate}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{The value to begin the estimate at.}

\item{delta}{The value to check against before stopping the estimate.}

\item{llik}{The log function to use to predict against.}

\item{xrange}{The range of the graph.}

\item{parameter}{The name of the parameter to put on the graph.}
}
\value{
A NR maximum likelihood graph.
}
\description{
Creates a NR maximum likelihood estimate
}
