% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate the number of tickets given the overbooking problem within the acceptable range}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of tickets sold}

\item{gamma}{The acceptable range of overbooking}

\item{p}{The probability of a ticket holder showing up}
}
\value{
A list containing the number of tickets for the discrete and continuous cases, the number of tickets sold, the probability of a ticket holder showing up, and the acceptable range of overbooking
}
\description{
Calculate the number of tickets given the overbooking problem within the acceptable range
}
