% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsamples.R
\name{plotsamples}
\alias{plotsamples}
\title{Plots a given number of iterations that have a size of choose(n)}
\usage{
plotsamples(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{The number of samples}

\item{iter}{The number of iterations}

\item{time}{The time to wait}
}
\value{
Plots that represent sample sizes
}
\description{
Plots a given number of iterations that have a size of choose(n)
}
